backend/requirements.txt
Flask==3.0.0
pymongo==4.6.1
mysql-connector-python==8.3.0
flask-cors==4.0.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==9.5.0